refactor(message)!: collapse the pool protocol onto a generic extension seam
`InterDaemonEvent`, four on `DaemonRequest`, two on `DaemonReply`, plus a
pool-named zenoh topic. Naming the protocol after one transport freezes
that transport's architecture into dora's 1.0 surface, which is exactly
what #1872 declined to do and what `docs/extensions.md` exists to avoid.

Replaces all of it with three opaque carriers:

  InterDaemonEvent::ExtensionMessage { dataflow_id, namespace,
                                       target_machine, payload }
  DaemonRequest::ExtensionRequest    { namespace, payload }
  DaemonReply::ExtensionReply        { payload }

and generalizes `dataflow_memory_pool_topic` to
`dataflow_extension_topic(dataflow_id, namespace)`.

The pool's own messages move into the extension, as
`dora_tensor_pool::protocol` — `NodeRequest`/`NodeResponse`/`PeerMessage`,
postcard-encoded into `payload`. dora reads exactly two things now: the
namespace, to pick an extension, and `target_machine`, to drop a
dataflow-scope broadcast on daemons it does not name. That machine gate
was previously duplicated in the `RegisterPool` and `FreePool` handlers;
it is one generic check in `handle_extension_message` now.

Node-side, four pool-shaped methods on `DoraNode` become one
`extension_request(namespace, payload)`. A second extension needs no
protocol change at all.

Two incidental fixes fall out of the rewrite:

  - the payload rides `bulk_bytes::vec`, so a cross-machine tensor frame
    is no longer serialized one `u8` at a time (the #3156 treatment, which
    `MemoryPoolWrite.tensor_data` never got).
  - `NodeRequest::Write` drops the `tensor_data` field. Every caller
    passed `Vec::new()` — the daemon reads the sender's segment itself —
    so the payload-carrying branch was unreachable.
