+++ apis/rust/node/src/daemon_connection/node_integration_testing.rs
@@ -140,13 +140,8 @@ impl IntegrationTestingEvents {
                 println!("{}", "node reports EventStreamDropped".blue());
                 DaemonReply::Result(Ok(()))
             }
-            DaemonRequest::FreePinnedMemory { .. } | DaemonRequest::WritePinnedMemory { .. } => {
-                DaemonReply::Result(Ok(()))
-            }
-            DaemonRequest::RegisterCrossMachinePool { .. } => {
-                eyre::bail!(
-                    "cross-machine pool registration is not supported in integration-testing mode"
-                )
+            DaemonRequest::ExtensionRequest { namespace, .. } => {
+                eyre::bail!("extension {namespace} is not available in integration-testing mode")
             }
             DaemonRequest::NodeConfig { .. } => {
                 eyre::bail!("unexpected NodeConfig in interactive mode")
