+++ apis/rust/node/src/node/control_channel.rs
@@ -211,84 +211,27 @@ impl ControlChannel {
         }
     }
 
-    pub fn write_pinned_memory(
+    /// Send an opaque request to the extension registered under
+    /// `namespace` on this node's daemon, and return its opaque reply.
+    ///
+    /// dora interprets neither side; see `docs/extensions.md`.
+    pub fn extension_request(
         &mut self,
-        shared_memory_id: String,
-        tensor_data: Vec<u8>,
-        size: usize,
-    ) -> eyre::Result<()> {
-        let request = DaemonRequest::WritePinnedMemory {
-            shared_memory_id,
-            tensor_data,
-            size,
-        };
+        namespace: String,
+        payload: Vec<u8>,
+    ) -> eyre::Result<Vec<u8>> {
+        let request = DaemonRequest::ExtensionRequest { namespace, payload };
         let reply = self
             .channel
             .request(&Timestamped {
                 inner: request,
                 timestamp: self.clock.new_timestamp(),
             })
-            .wrap_err("failed to send WritePinnedMemory request to dora-daemon")?;
+            .wrap_err("failed to send ExtensionRequest to dora-daemon")?;
         match reply {
-            DaemonReply::Result(Ok(())) => Ok(()),
+            DaemonReply::ExtensionReply { payload } => Ok(payload),
             DaemonReply::Result(Err(e)) => bail!("{e}"),
-            other => bail!("unexpected WritePinnedMemory reply: {other:?}"),
-        }
-    }
-
-    /// Register a pool on a remote machine via the daemon (the daemon
-    /// resolves the machine through the coordinator and mirrors the
-    /// pool there with a synchronous confirmation).
-    #[allow(clippy::too_many_arguments)]
-    pub fn register_cross_machine_pool(
-        &mut self,
-        shared_memory_id: String,
-        shmem_name: String,
-        size: usize,
-        dtype: String,
-        shape: Vec<i64>,
-        device: String,
-        machine_id: String,
-    ) -> eyre::Result<(Result<(), String>, bool)> {
-        let request = DaemonRequest::RegisterCrossMachinePool {
-            shared_memory_id,
-            shmem_name,
-            size,
-            dtype,
-            shape,
-            device,
-            machine_id,
-        };
-        let reply = self
-            .channel
-            .request(&Timestamped {
-                inner: request,
-                timestamp: self.clock.new_timestamp(),
-            })
-            .wrap_err("failed to send RegisterCrossMachinePool request to dora-daemon")?;
-        match reply {
-            DaemonReply::CrossMachinePoolRegistered { result, direct } => Ok((result, direct)),
-            other => bail!("unexpected RegisterCrossMachinePool reply: {other:?}"),
-        }
-    }
-
-    /// Release a memory pool through the daemon. Cross-machine pools get
-    /// their tracking entry removed, the peer is told (targeted `FreePool`
-    /// publish), and the local mirror is unlinked; local-only pools are a
-    /// plain acknowledgement (the pool's own cleanup is node-side).
-    pub fn free_pinned_memory(&mut self, shared_memory_id: String) -> eyre::Result<()> {
-        let request = DaemonRequest::FreePinnedMemory { shared_memory_id };
-        let reply = self
-            .channel
-            .request(&Timestamped {
-                inner: request,
-                timestamp: self.clock.new_timestamp(),
-            })
-            .wrap_err("failed to send FreePinnedMemory request to dora-daemon")?;
-        match reply {
-            DaemonReply::Result(Ok(())) => Ok(()),
-            DaemonReply::Result(Err(e)) => bail!("{e}"),
-            other => bail!("unexpected FreePinnedMemory reply: {other:?}"),
+            other => bail!("unexpected ExtensionRequest reply: {other:?}"),
         }
     }
 }
