+++ binaries/daemon/src/memory_pool/auth.rs
@@ -14,7 +14,7 @@ use super::*;
 ///
 /// Scope: the token authenticates the direct-TCP data plane only. The
 /// zenoh control/relay plane (RegisterPool/FreePool/MemoryPoolWrite over
-/// `dataflow_memory_pool_topic`) is unauthenticated — dora's mesh-daemon
+/// `dataflow_extension_topic`) is unauthenticated — dora's mesh-daemon
 /// model treats peers as trusted. When this token is relied on over an
 /// untrusted link (edge↔cloud WAN), the zenoh transport itself must be
 /// secured (e.g. TLS via the zenoh config), or the control plane must be
