+++ binaries/daemon/src/node_communication/mod.rs
@@ -245,32 +245,6 @@ impl Listener {
                     .await
                     .wrap_err("failed to send register reply")?;
             }
-            DaemonRequest::RegisterCrossMachinePool {
-                shared_memory_id,
-                shmem_name,
-                size,
-                dtype,
-                shape,
-                device,
-                machine_id,
-            } => {
-                let (reply_sender, reply) = oneshot::channel();
-                self.process_daemon_event(
-                    DaemonNodeEvent::RegisterCrossMachinePool {
-                        shared_memory_id,
-                        shmem_name,
-                        size,
-                        dtype,
-                        shape,
-                        device,
-                        machine_id,
-                        reply_sender,
-                    },
-                    Some(reply),
-                    connection,
-                )
-                .await?;
-            }
             DaemonRequest::NodeConfig { .. } => {
                 let reply = DaemonReply::Result(Err("unexpected node config message".into()));
                 self.send_reply(reply, connection)
@@ -424,29 +398,12 @@ impl Listener {
                 )
                 .await?;
             }
-            DaemonRequest::WritePinnedMemory {
-                shared_memory_id,
-                tensor_data,
-                size,
-            } => {
+            DaemonRequest::ExtensionRequest { namespace, payload } => {
                 let (reply_sender, reply) = oneshot::channel();
                 self.process_daemon_event(
-                    DaemonNodeEvent::WriteMemoryPool {
-                        shared_memory_id,
-                        tensor_data,
-                        size,
-                        reply_sender,
-                    },
-                    Some(reply),
-                    connection,
-                )
-                .await?;
-            }
-            DaemonRequest::FreePinnedMemory { shared_memory_id } => {
-                let (reply_sender, reply) = oneshot::channel();
-                self.process_daemon_event(
-                    DaemonNodeEvent::FreePinnedMemory {
-                        shared_memory_id,
+                    DaemonNodeEvent::ExtensionRequest {
+                        namespace,
+                        payload,
                         reply_sender,
                     },
                     Some(reply),
