+++ libraries/extensions/tensor-pool/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod protocol;
+
 use std::collections::{HashMap, HashSet};
 use std::sync::{Arc, Mutex};
