+++ libraries/message/src/daemon_to_node.rs
@@ -101,18 +101,15 @@ pub enum DaemonReply {
         value: Option<Vec<u8>>,
     },
     Empty,
-    /// Result of a cross-machine pool registration. `Err` carries the
-    /// warning message (resolution failure or remote creation failure) —
-    /// the register is a warn-and-no-op in both cases. `direct` tells the
-    /// node whether the remote daemon can open its segment directly
-    /// (same host): when true, the per-frame data push is skipped.
+    /// Opaque reply to [`crate::node_to_daemon::DaemonRequest::ExtensionRequest`],
+    /// produced by the extension's daemon half. dora does not interpret it.
     ///
-    /// Appended last so existing variants keep their bincode indices: the
+    /// Appended last so existing variants keep their postcard indices: the
     /// Python node API ships separately (PyPI) from the daemon, so a
     /// mixed-version pair must not misdecode older replies.
-    CrossMachinePoolRegistered {
-        result: Result<(), String>,
-        direct: bool,
+    ExtensionReply {
+        #[serde(with = "crate::bulk_bytes::vec")]
+        payload: Vec<u8>,
     },
 }
 
@@ -134,7 +131,7 @@ impl DaemonReply {
             // its own length is the whole hint.
             DaemonReply::ExtensionValue { value } => value.as_ref().map_or(0, |bytes| bytes.len()),
             DaemonReply::Result(_) | DaemonReply::NodeConfig { .. } | DaemonReply::Empty => 0,
-            DaemonReply::CrossMachinePoolRegistered { .. } => 0,
+            DaemonReply::ExtensionReply { payload } => payload.len(),
         }
     }
 }
